Module ssd1306::mode

Operating modes for the SSD1306

This driver can be used in different modes. A mode defines how the driver will behave, and what methods it exposes. Look at the modes below for more information on what they expose.

Re-exports

pub use self::graphics::GraphicsMode;

pub use self::raw::RawMode;

pub use self::terminal::TerminalMode;

Modules

displaymode
Abstraction of different operating modes for the SSD1306

graphics
Buffered display module for use with the embedded_graphics crate

raw
Raw mode for coercion into richer driver types

terminal
Unbuffered terminal display mode