Struct ssd1306::mode::graphics::GraphicsMode

pub struct GraphicsMode<DI> where
    DI: DisplayInterface,  { /* fields omitted */ }

Graphics mode handler

Methods

impl<DI> GraphicsMode<DI> where
DI: DisplayInterface,

pub fn clear(&mut self)

Clear the display buffer. You need to call disp.flush() for any effect on the screen

pub fn reset<RST, DELAY>(&mut self, rst: &mut RST, delay: &mut DELAY) where
RST: OutputPin,
DELAY: DelayMs<u8>,

Reset display

pub fn flush(&mut self) -> Result<(), ()>

Write out data to display

pub fn set_pixel(&mut self, x: u32, y: u32, value: u8)

Turn a pixel on or off. A non-zero value is treated as on, 0 as off. If the X and Y coordinates are out of the bounds of the display, this method call is a noop.

pub fn init(&mut self) -> Result<(), ()>

Display is set up in column mode, i.e. a byte walks down a column of 8 pixels from column 0 on the left, to column n on the right

pub fn get_dimensions(&self) -> (u8, u8)

Get display dimensions, taking into account the current rotation of the display

pub fn set_rotation(&mut self, rot: DisplayRotation) -> Result<(), ()>

Set the display rotation