ssd1306::mode

Trait TerminalDisplaySize

pub trait TerminalDisplaySize: DisplaySize {
    const CHAR_NUM: u8;
}

Extends the DisplaySize trait to include number of characters that can fit on the display.

Required Associated Constants

const CHAR_NUM: u8

The number of characters that can fit on the display at once (w * h / (8 * 8))

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.