ssd1306

Module prelude

Crate prelude

Re-exports

pub use super::mode::DisplayConfig;

pub use super::rotation::DisplayRotation;

pub use super::size::DisplaySize;

pub use super::size::DisplaySize128x32;

pub use super::size::DisplaySize128x64;

pub use super::size::DisplaySize64x48;

pub use super::size::DisplaySize72x40;

pub use super::size::DisplaySize96x16;

Structs

Brightness

Struct that holds display brightness

I2CInterface

I2C communication interface

SPIInterface

SPI display interface.

Traits

WriteOnlyDataCommand

This trait implements a write-only interface for a display which has separate data and command modes. It is the responsibility of implementations to activate the correct mode in their implementation when corresponding method is called.