ssd1306::size

Trait NewZeroed

pub trait NewZeroed {
    // Required method
    fn new_zeroed() -> Self;
}

Workaround trait, since Default is only implemented to arrays up to 32 of size

Required Methods

fn new_zeroed() -> Self

Creates a new value with its memory set to zero

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.