ssd1306

Struct Ssd1306

pub struct Ssd1306<DI, SIZE, MODE> { /* private fields */ }

SSD1306 driver.

Note that some methods are only available when the display is configured in a certain mode.

Implementations

impl<DI, SIZE> Ssd1306<DI, SIZE, BufferedGraphicsMode<SIZE>>
where DI: WriteOnlyDataCommand, SIZE: DisplaySize,

pub fn clear_buffer(&mut self)

Clear the underlying framebuffer. You need to call disp.flush() for any effect on the screen.

pub fn flush(&mut self) -> Result<(), DisplayError>

Write out data to a display.

This only updates the parts of the display that have changed since the last flush.

pub fn set_pixel(&mut self, x: u32, y: u32, value: bool)

Turn a pixel on or off. A non-zero value is treated as on, 0 as off. If the X and Y coordinates are out of the bounds of the display, this method call is a noop.

impl<DI, SIZE> Ssd1306<DI, SIZE, TerminalMode>
where DI: WriteOnlyDataCommand, SIZE: TerminalDisplaySize,

pub fn clear(&mut self) -> Result<(), TerminalModeError>

Clear the display and reset the cursor to the top left corner

pub fn print_char(&mut self, c: char) -> Result<(), TerminalModeError>

Print a character to the display

pub fn position(&self) -> Result<(u8, u8), TerminalModeError>

Get the current cursor position, in character coordinates. This is the (column, row) that the next character will be written to.

pub fn set_position( &mut self, column: u8, row: u8, ) -> Result<(), TerminalModeError>

Set the cursor position, in character coordinates. This is the (column, row) that the next character will be written to. If the position is out of bounds, an Err will be returned.

impl<DI, SIZE> Ssd1306<DI, SIZE, BasicMode>
where DI: WriteOnlyDataCommand, SIZE: DisplaySize,

pub fn clear(&mut self) -> Result<(), DisplayError>

Clear the display.

impl<DI, SIZE> Ssd1306<DI, SIZE, BasicMode>
where SIZE: DisplaySize,

pub fn new(interface: DI, size: SIZE, rotation: DisplayRotation) -> Self

Create a basic SSD1306 interface.

Use the into_*_mode methods to enable more functionality.

impl<DI, SIZE, MODE> Ssd1306<DI, SIZE, MODE>
where SIZE: DisplaySize,

pub fn into_buffered_graphics_mode( self, ) -> Ssd1306<DI, SIZE, BufferedGraphicsMode<SIZE>>

Convert the display into a buffered graphics mode, supporting embedded-graphics.

See BufferedGraphicsMode for more information.

pub fn into_terminal_mode(self) -> Ssd1306<DI, SIZE, TerminalMode>

Convert the display into a text-only, terminal-like mode.

See TerminalMode for more information.

impl<DI, SIZE, MODE> Ssd1306<DI, SIZE, MODE>
where DI: WriteOnlyDataCommand, SIZE: DisplaySize,

pub fn init_with_addr_mode( &mut self, mode: AddrMode, ) -> Result<(), DisplayError>

Initialise the display in one of the available addressing modes.

pub fn set_addr_mode(&mut self, mode: AddrMode) -> Result<(), DisplayError>

Change the addressing mode

pub fn bounded_draw( &mut self, buffer: &[u8], disp_width: usize, upper_left: (u8, u8), lower_right: (u8, u8), ) -> Result<(), DisplayError>

Send the data to the display for drawing at the current position in the framebuffer and advance the position accordingly. Cf. set_draw_area to modify the affected area by this method.

This method takes advantage of a bounding box for faster writes.

pub fn draw(&mut self, buffer: &[u8]) -> Result<(), DisplayError>

Send a raw buffer to the display.

pub fn dimensions(&self) -> (u8, u8)

Get display dimensions, taking into account the current rotation of the display

use ssd1306::{mode::TerminalMode, prelude::*, Ssd1306};

let mut display = Ssd1306::new(
    interface,
    DisplaySize128x64,
    DisplayRotation::Rotate0,
).into_terminal_mode();
assert_eq!(display.dimensions(), (128, 64));

let mut rotated_display = Ssd1306::new(
    interface,
    DisplaySize128x64,
    DisplayRotation::Rotate90,
).into_terminal_mode();
assert_eq!(rotated_display.dimensions(), (64, 128));

pub fn rotation(&self) -> DisplayRotation

Get the display rotation.

pub fn set_rotation( &mut self, rotation: DisplayRotation, ) -> Result<(), DisplayError>

Set the display rotation.

pub fn set_mirror(&mut self, mirror: bool) -> Result<(), DisplayError>

Set mirror enabled/disabled.

pub fn set_brightness( &mut self, brightness: Brightness, ) -> Result<(), DisplayError>

Change the display brightness.

pub fn set_display_on(&mut self, on: bool) -> Result<(), DisplayError>

Turn the display on or off. The display can be drawn to and retains all of its memory even while off.

pub fn set_draw_area( &mut self, start: (u8, u8), end: (u8, u8), ) -> Result<(), DisplayError>

Set the position in the framebuffer of the display limiting where any sent data should be drawn. This method can be used for changing the affected area on the screen as well as (re-)setting the start point of the next draw call.

pub fn set_column(&mut self, column: u8) -> Result<(), DisplayError>

Set the column address in the framebuffer of the display where any sent data should be drawn.

pub fn set_row(&mut self, row: u8) -> Result<(), DisplayError>

Set the page address (row 8px high) in the framebuffer of the display where any sent data should be drawn.

Note that the parameter is in pixels, but the page will be set to the start of the 8px row which contains the passed-in row.

pub fn set_invert(&mut self, invert: bool) -> Result<(), DisplayError>

Set the screen pixel on/off inversion

pub fn release(self) -> DI

Release the contained interface.

impl<DI, SIZE, MODE> Ssd1306<DI, SIZE, MODE>

pub fn reset<RST, DELAY>( &mut self, rst: &mut RST, delay: &mut DELAY, ) -> Result<(), Error<Infallible, RST::Error>>
where RST: OutputPin, DELAY: DelayNs,

Reset the display.