Struct ssd1306::properties::DisplayProperties

pub struct DisplayProperties<DI> { /* fields omitted */ }

Display properties struct

Methods

impl<DI> DisplayProperties<DI> where
DI: DisplayInterface,

pub fn new(
iface: DI,
display_size: DisplaySize,
display_rotation: DisplayRotation
) -> DisplayProperties<DI>

Create new DisplayProperties instance

pub fn init_column_mode(&mut self) -> Result<(), ()>

Initialise the display in column mode (i.e. a byte walks down a column of 8 pixels) with column 0 on the left and column (display_width - 1) on the right.

pub fn set_draw_area(
&mut self,
start: (u8, u8),
end: (u8, u8)
) -> Result<(), ()>

Set the position in the framebuffer of the display where any sent data should be drawn. This method can be used for changing the affected area on the screen as well as (re-)setting the start point of the next draw call.

pub fn draw(&mut self, buffer: &[u8]) -> Result<(), ()>

Send the data to the display for drawing at the current position in the framebuffer and advance the position accordingly. Cf. set_draw_area to modify the affected area by this method.

pub fn get_size(&self) -> DisplaySize

Get the configured display size

pub fn get_dimensions(&self) -> (u8, u8)

Get display dimensions, taking into account the current rotation of the display

let disp = DisplayProperties::new(
    interface,
    DisplaySize::Display128x64,
    DisplayRotation::Rotate0
);
assert_eq!(disp.get_dimensions(), (128, 64));

let rotated_disp = DisplayProperties::new(
    interface,
    DisplaySize::Display128x64,
    DisplayRotation::Rotate90
);
assert_eq!(rotated_disp.get_dimensions(), (64, 128));

pub fn get_rotation(&self) -> DisplayRotation

Get the display rotation

pub fn set_rotation(
&mut self,
display_rotation: DisplayRotation
) -> Result<(), ()>

Set the display rotation